pyproject.toml
[tool.poetry]
name = "fastgql"
packages = [ {include = "fastgql"} ]
version = "0.1.1"
description = "The easiest, fastest python GraphQL framework."
authors = ["Jeremy Berman <jerber@sas.upenn.edu>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/jerber/fastgql"
repository = "https://github.com/jerber/fastgql"
documentation = "https://jerber.github.io/fastgql/"
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Database :: Database Engines/Servers",
    "Topic :: Internet",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.4.2"
fastapi = "^0.104.0"
edgedb = { version = "^1.7.0", optional = true }
graphql-core = "^3.2.3"
devtools = "^0.12.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}

[tool.poetry.extras]
edgedb = ["edgedb"]

[tool.poetry.dev-dependencies]
ruff = '*'
mypy = '^1.5'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"